Per https://signalfuse.atlassian.net/browse/FUSE-7278 I enhanced the collectd cpu plug-in to generate the new metric "cpu.utilization", which is the number of active (i.e. non-idle) CPU cycles.
This metric is not generated by default. To configure the CPU plug-in to generate the metric the configuration option "ReportUtilizedState" must be set to true in the CPU plug-in's configuration block in collectd.conf. The CPU plug-in must also be configured to report by individual CPU states, otherwise the plug-in's existing legacy behavior is retained - specifically, it only reports a single "percent.active" metric for the host.